Added Plugin Schema Verification
