+++ res/plugin_schema.xsd
@@ -5,53 +5,19 @@
             <xs:sequence>
                 <xs:element name="name" type="xs:string"/>
                 <xs:element name="description" type="xs:string"/>
-                <xs:element name="pois">
+                <xs:element name="pois" maxOccurs="unbounded">
                     <xs:complexType>
                         <xs:sequence>
-                            <xs:element name="dll">
-                                <xs:complexType>
-                                    <xs:sequence>
-                                        <xs:element name="name" type="xs:string"/>
-                                    </xs:sequence>
-                                </xs:complexType>
-                            </xs:element>
-                            <xs:element name="variable">
-                                <xs:complexType>
-                                    <xs:sequence>
-                                        <xs:element name="type" type="xs:string"/>
-                                        <xs:element name="value" type="xs:string"/>
-                                        <xs:element name="address" type="xs:string"/>
-                                    </xs:sequence>
-                                </xs:complexType>
-                            </xs:element>
-                            <xs:element name="string">
-                                <xs:complexType>
-                                    <xs:sequence>
-                                        <xs:element name="value" type="xs:string"/>
-                                        <xs:element name="address" type="xs:string"/>
-                                    </xs:sequence>
-                                </xs:complexType>
-                            </xs:element>
-                            <xs:element name="function">
-                                <xs:complexType>
-                                    <xs:sequence>
-                                        <xs:element name="paramOrder" type="xs:string"/>
-                                        <xs:element name="paramType" type="xs:string"/>
-                                        <xs:element name="paramValue" type="xs:string"/>
-                                        <xs:element name="returnType" type="xs:string"/>
-                                        <xs:element name="returnValue" type="xs:string"/>
-                                    </xs:sequence>
-                                </xs:complexType>
-                            </xs:element><xs:element name="struct">
-                                <xs:complexType>
-                                    <xs:sequence>
-                                        <xs:element name="address" type="xs:string"/>
-                                        <xs:element name="order" type="xs:string"/>
-                                        <xs:element name="type" type="xs:string"/>
-                                        <xs:element name="value" type="xs:string"/>
-                                    </xs:sequence>
-                                </xs:complexType>
-                            </xs:element>
+                            <xs:element name="name" type="xs:string"/>
+                            <xs:element name="type" type="xs:string"/>
+                            <xs:element name="mapping" type="xs:string"/>
+                        </xs:sequence>
+                    </xs:complexType>
+                </xs:element>
+                <xs:element name="types">
+                    <xs:complexType>
+                        <xs:sequence maxOccurs="unbounded">
+                            <xs:element name="name" type="xs:string"/>
                         </xs:sequence>
                     </xs:complexType>
                 </xs:element>
