+++ src/common/tab_layout.py
@@ -1,6 +1,6 @@
 from PyQt5.QtCore import Qt
 from PyQt5.QtGui import QFont, QWheelEvent
-from PyQt5.QtWidgets import QWidget, QGridLayout, QFrame, QLabel, QVBoxLayout, QFileDialog
+from PyQt5.QtWidgets import QWidget, QGridLayout, QFrame, QLabel, QVBoxLayout, QFileDialog, QGroupBox
 
 from src.common import constants
 
@@ -9,9 +9,9 @@ class TabLayout(QWidget):
     def __init__(self, leftPanelLabel, rightPanelLabel, hasTopPanel=False):
         super().__init__()
         self.__mainGrid = QGridLayout()
-        leftFrame = QFrame()
-        rightFrame = QFrame()
-        TopFrame = QFrame()
+        leftFrame = QGroupBox(leftPanelLabel)
+        rightFrame = QGroupBox(rightPanelLabel)
+        TopFrame = QGroupBox()
         self.setFont(QFont("arial", 11))
 
         self._leftPanelLayout = QVBoxLayout()
@@ -21,28 +21,6 @@ class TabLayout(QWidget):
         leftFrame.setLayout(self._leftPanelLayout)
         rightFrame.setLayout(self._rightPanelLayout)
         TopFrame.setLayout(self._TopPanelLayout)
-
-        self.__leftPanelLabel = QLabel()
-        self.__rightPanelLabel = QLabel()
-        self.__TopPanelLabel = QLabel()
-
-        self.__leftPanelLabel.setStyleSheet(constants.PANEL_TITLE_STYLE)
-        self.__rightPanelLabel.setStyleSheet(constants.PANEL_TITLE_STYLE)
-
-        self.__leftPanelLabel.setText(leftPanelLabel)
-        self.__rightPanelLabel.setText(rightPanelLabel)
-
-        self.__leftPanelLabel.setAlignment(Qt.AlignCenter)
-        self.__rightPanelLabel.setAlignment(Qt.AlignCenter)
-        self.__TopPanelLabel.setAlignment(Qt.AlignCenter)
-
-        self._leftPanelLayout.addWidget(self.__leftPanelLabel)
-        self._rightPanelLayout.addWidget(self.__rightPanelLabel)
-        self._TopPanelLayout.addWidget(self.__TopPanelLabel)
-
-        self._leftPanelLayout.setAlignment(self.__leftPanelLabel, Qt.AlignTop)
-        self._rightPanelLayout.setAlignment(self.__rightPanelLabel, Qt.AlignTop)
-        self._TopPanelLayout.setAlignment(self.__TopPanelLabel, Qt.AlignTop)
 
         if hasTopPanel:
             self.__mainGrid.addWidget(TopFrame, 0, 0, 1, 9)
