+++ src/items/poi_item.py
@@ -1,4 +1,7 @@
 class PoiItem:
-    def __init__(self, type):
-        self.data = dict()
-        self.data['type'] = type
+    def __init__(self, type, data=None):
+        if data is None:
+            self.data = {}
+            self.data['type'] = type
+        else:
+            self.data = data
